+++ src/mln/renderer/property_evaluator.hpp
@@ -17,7 +17,11 @@ public:
 
     T operator()(const style::Undefined&) const { return defaultValue; }
     T operator()(const T& constant) const { return constant; }
-    T operator()(const style::PropertyExpression<T>& fn) const { return fn.evaluate(parameters.z); }
+    T operator()(const style::PropertyExpression<T>& fn) const {
+        return fn.evaluate(
+            style::expression::EvaluationContext(parameters.z).withGlobalState(parameters.globalState.get()),
+            defaultValue);
+    }
 
 private:
     const PropertyEvaluationParameters& parameters;