+++ src/mln/renderer/sources/render_image_source.cpp
@@ -68,11 +68,12 @@ std::unordered_map<std::string, std::vector<Feature>> RenderImageSource::queryRe
     const TransformState&,
     const std::unordered_map<std::string, const RenderLayer*>&,
     const RenderedQueryOptions&,
+    const GlobalStateMap*,
     const mat4&) const {
     return std::unordered_map<std::string, std::vector<Feature>>{};
 }
 
-std::vector<Feature> RenderImageSource::querySourceFeatures(const SourceQueryOptions&) const {
+std::vector<Feature> RenderImageSource::querySourceFeatures(const SourceQueryOptions&, const GlobalStateMap*) const {
     return {};
 }