+++ src/mln/style/layers/background_layer_properties.cpp
@@ -35,6 +35,14 @@ expression::Dependency BackgroundLayerProperties::getDependencies() const noexce
     return layerImpl().paint.getDependencies();
 }
 
+expression::Dependency BackgroundLayerProperties::getEvaluatedDependencies() const noexcept {
+    return evaluated.getDependencies();
+}
+
+void BackgroundLayerProperties::collectEvaluatedGlobalStateRefs(std::set<std::string>& refs) const {
+    evaluated.collectGlobalStateRefs(refs);
+}
+
 } // namespace style
 } // namespace mln