+++ src/mln/style/layers/circle_layer_properties.cpp
@@ -33,6 +33,14 @@ expression::Dependency CircleLayerProperties::getDependencies() const noexcept {
     return layerImpl().paint.getDependencies() | layerImpl().layout.getDependencies();
 }
 
+expression::Dependency CircleLayerProperties::getEvaluatedDependencies() const noexcept {
+    return evaluated.getDependencies();
+}
+
+void CircleLayerProperties::collectEvaluatedGlobalStateRefs(std::set<std::string>& refs) const {
+    evaluated.collectGlobalStateRefs(refs);
+}
+
 } // namespace style
 } // namespace mln