+++ src/mln/style/layers/fill_extrusion_layer_impl.hpp
@@ -14,6 +14,15 @@ public:
     bool hasLayoutDifference(const Layer::Impl&) const override;
     void stringifyLayout(rapidjson::Writer<rapidjson::StringBuffer>&) const override;
 
+    expression::Dependency getLayoutDependencies() const noexcept override {
+        return layout.getDependencies() | Layer::Impl::getLayoutDependencies();
+    }
+
+    void collectLayoutGlobalStateRefs(std::set<std::string>& refs) const override {
+        layout.collectGlobalStateRefs(refs);
+        Layer::Impl::collectLayoutGlobalStateRefs(refs);
+    }
+
     FillExtrusionLayoutProperties::Unevaluated layout;
     FillExtrusionPaintProperties::Transitionable paint;