+++ src/mln/style/parser.hpp
@@ -8,6 +8,7 @@
 #include <mln/text/glyph.hpp>
 
 #include <mln/util/constants.hpp>
+#include <mln/util/feature.hpp>
 #include <mln/util/rapidjson.hpp>
 #include <mln/util/font_stack.hpp>
 #include <mln/util/geo.hpp>
@@ -40,6 +41,10 @@ public:
     TransitionOptions transition{{util::DEFAULT_TRANSITION_DURATION}};
     Light light;
 
+    // Default values of the root "state" property, keyed by state property
+    // name, used by the "global-state" expression.
+    GlobalStateMap globalStateDefaults;
+
     std::string name;
     LatLng latLng;
     double zoom = 0;
@@ -54,6 +59,7 @@ public:
 private:
     void parseTransition(const JSValue&);
     void parseLight(const JSValue&);
+    void parseState(const JSValue&);
     void parseSources(const JSValue&);
     void parseSprites(const JSValue&);
     void parseLayers(const JSValue&);